Paragraph Summary/requirements.txt
streamlit
transformers
bert-extractive-summarizer
ipywidgets
tqdm
sentencepiece